MXS-2253 Make Cache use the new duration config type
Internally the Cache used seconds, so some changes were needed.

+++ server/modules/filter/cache/storage/storage_inmemory/inmemorystorage.hh
@@ -73,8 +73,8 @@ private:
         {
         }
 
-        uint32_t time;
-        Value    value;
+        int64_t time;
+        Value   value;
     };
 
     struct Stats
